+++ src/code/error/all_error.rs
@@ -1,4 +1,4 @@
-// devela::code::result::error::all_error
+// devela::code::error::all_error
 //
 //! Defines the [`AllError`] enum.
 //