+++ src/code/error/reexports.rs
@@ -1,4 +1,4 @@
-// devela::code::result::error::reexports
+// devela::code::error::reexports
 //
 //! Reexported error-related items.
 //
@@ -9,6 +9,20 @@ use crate::reexport;
 reexport! { rust: core::error,
     doc: "A trait representing the basic expectations for error values.",
     Error
+}
+
+reexport! { rust: core,
+    doc: "Causes compilation to fail with the given error message when encountered.",
+    compile_error
+}
+
+reexport! { rust: std::backtrace,
+    doc: "A captured OS thread stack backtrace.",
+    Backtrace
+}
+reexport! { rust: std::backtrace,
+    doc: "The current status of a backtrace.",
+    BacktraceStatus
 }
 
 #[cfg(feature = "error")]