+++ src/code/mod.rs
@@ -1,10 +1,11 @@
 // devela::code
 //
 //! Code reflective synthesis.
-#![doc = crate::doc_!(modules: crate; code: marker, ops, result, util)]
+#![doc = crate::doc_!(modules: crate; code: error, marker, ops, panic, result, util)]
 #![doc = crate::doc_!(newline)]
 //!
-#![doc = crate::doc_!(extends: any, clone, convert, default, hint, marker, ops)]
+#![doc = crate::doc_!(extends:
+    any, clone, convert, default, error, hint, marker, ops, panic, result)]
 //
 // safety
 #![cfg_attr(feature = "safe_code", forbid(unsafe_code))]
@@ -14,8 +15,10 @@ mod default; // ConstDefault, Default
 mod guard; // ScopeGuard
 mod reexports; // re-exported items
 
+pub mod error; // AllError, modular errors
 pub mod marker; // core::marker, type_marker!, type_resource!, TypeResource, TypeResourced
-pub mod ops; // re-exported overloadable operators
+pub mod ops; // ::core::ops::*
+pub mod panic; // Panic, define_panic_handler!
 pub mod result; // AllError, serr, sok, Mismatch, OptRes, ValueQuant…
 pub mod util; // utility macros and functions
 
@@ -29,7 +32,9 @@ crate::items! { // structural access: _mods, _pub_mods, _internals, _all, _alway
         pub use super::{any::_all::*, default::*, guard::*, reexports::*};
     }
     mod _pub_mods {
-        pub use super::{marker::_all::*, ops::_all::*, result::_all::*, util::_all::*};
+        pub use super::{
+            error::_all::*, marker::_all::*, ops::_all::*, result::_all::*, util::_all::*,
+        };
     }
     pub(super) mod _internals {
         pub(crate) use super::util::_internals::*;
@@ -39,6 +44,8 @@ crate::items! { // structural access: _mods, _pub_mods, _internals, _all, _alway
         pub use super::{_pub_mods::*, _mods::*};
     }
     pub(super) mod _always { #![allow(unused)]
-        pub use super::{marker::_always::*, reexports::*, result::_always::*, util::_always::*};
+        pub use super::{
+            error::_always::*, marker::_always::*, panic::_always::*,
+            reexports::*, result::_always::*, util::_always::*, };
     }
 }