+++ src/code/panic/namespace.rs
@@ -1,4 +1,4 @@
-// devela::code::result::panic::namespace
+// devela::code::panic::namespace
 //
 //! Defines the [`Panic`] namespace.
 //